+++ shiny/experimental/ui/__init__.py
@@ -1,10 +1,7 @@
 # Experimental
 
-from ._card import (
-    card_image,
-    card_title,
-)
-from ._deprecated import card, card_body
+from ._card import card_image
+from ._deprecated import card, card_body, card_title
 
 __all__ = (
     # Deprecated
