+++ shiny/experimental/ui/_card.py
@@ -6,69 +6,16 @@ import mimetypes
 from pathlib import Path, PurePath
 from typing import Literal, Optional, Protocol
 
-from htmltools import (
-    Tag,
-    TagAttrs,
-    TagAttrValue,
-    TagChild,
-    TagFunction,
-    Tagifiable,
-    css,
-    tags,
-)
+from htmltools import Tag, TagAttrs, TagAttrValue, Tagifiable, css, tags
 
 from ..._docstring import add_example
 from ...ui._card import CardItem, card_body
 from ...ui.css import CssUnit, as_css_unit
 from ...ui.fill import as_fill_item, as_fillable_container
 
-__all__ = (
-    # Questioning:
-    "card_title",
-    "card_image",
-)
+__all__ = ("card_image",)
 
 ############################################################################
-
-
-@add_example()
-def card_title(
-    *args: TagChild | TagAttrs,
-    container: TagFunction = tags.h5,
-    **kwargs: TagAttrValue,
-) -> Tagifiable:
-    """
-    A card title container
-
-    :func:`~shiny.experimental.ui.card_title` creates a general container for the "title" of
-    a :func:`~shiny.ui.card`. This component is designed
-    to be provided as a direct child to :func:`~shiny.ui.card`.
-
-    Parameters
-    ----------
-    *args
-        Contents to appear in the card's title, or tag attributes to pass to the
-        resolved :class:`~htmltools.Tag` object.
-    container
-        Method for the returned :class:`~htmltools.Tag` object. Defaults to
-        :func:`~shiny.ui.tags.h5`.
-    **kwargs
-        Additional HTML attributes for the returned :class:`~htmltools.Tag` object.
-
-    Returns
-    -------
-    :
-        An :class:`~htmltools.Tag` object.
-
-    See Also
-    --------
-    * :func:`~shiny.ui.card` for creating a card component.
-    * :func:`~shiny.ui.card_header` for creating a header within a card.
-    * :func:`~shiny.experimental.ui.card_body` for putting content inside a card.
-    * :func:`~shiny.ui.card_footer` for creating a footer within a card.
-    * :func:`~shiny.experimental.ui.card_image` for adding an image to a card.
-    """
-    return container(*args, **kwargs)
 
 
 class ImgContainer(Protocol):
